oprava pozice a pridani dat

+++ Pozice.java
@@ -1,25 +1,25 @@
 public class Pozice {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Pozice(int x, int y){
+    public Pozice(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
